refactor(rpc): move cached Ethereum gas price fetcher to a new module
